Refactor test utility methods into a separate file
